+++ robocorp-code/codegen/settings.py
@@ -76,20 +76,26 @@ SETTINGS = [
     ),
     Setting(
         "robocorp.verifyLSP",
-        "true",
+        True,
         "Verify if the Robot Framework Language Server is installed?",
         setting_type="boolean",
     ),
     Setting(
         "robocorp.autoSetPythonExtensionInterpreter",
-        "true",
+        True,
         "If a file in a Robot is opened, the python extension interpreter is automatically set to match the Robot interpreter.",
         setting_type="boolean",
     ),
     Setting(
         "robocorp.autoSetPythonExtensionDisableActivateTerminal",
-        "true",
+        True,
         'Automatically sets the value of "python.terminal.activateEnvironment" to false to avoid wrong auto-activation when Robot terminal is created.',
+        setting_type="boolean",
+    ),
+    Setting(
+        "robocorp.proceedWithLongPathsDisabled",
+        False,
+        "Enables Robocorp Code to be started even with long paths disabled.",
         setting_type="boolean",
     ),
 ]