+++ robocorp-code/src/robocorp_code/rcc.py
@@ -75,7 +75,7 @@ def download_rcc(location: str, force: bool = False) -> None:
                     else:
                         relative_path = "/linux32/rcc"
 
-                RCC_VERSION = "v11.33.2"
+                RCC_VERSION = "v11.36.0"
                 prefix = f"https://downloads.robocorp.com/rcc/releases/{RCC_VERSION}"
                 url = prefix + relative_path