+++ robocorp-code/vscode-client/src/robocorpSettings.ts
@@ -19,6 +19,7 @@ export const ROBOCORP_HOME = "robocorp.home";
 export const ROBOCORP_VERIFY_LSP = "robocorp.verifyLSP";
 export const ROBOCORP_AUTO_SET_PYTHON_EXTENSION_INTERPRETER = "robocorp.autoSetPythonExtensionInterpreter";
 export const ROBOCORP_AUTO_SET_PYTHON_EXTENSION_DISABLE_ACTIVATE_TERMINAL = "robocorp.autoSetPythonExtensionDisableActivateTerminal";
+export const ROBOCORP_PROCEED_WITH_LONG_PATHS_DISABLED = "robocorp.proceedWithLongPathsDisabled";
 
 export function getLanguageServerTcpPort(): number {
     let key = ROBOCORP_LANGUAGE_SERVER_TCP_PORT;
@@ -163,6 +164,21 @@ export function getAutosetpythonextensiondisableactivateterminal(): boolean {
 
 export async function setAutosetpythonextensiondisableactivateterminal(value): Promise<void> {
     let key = ROBOCORP_AUTO_SET_PYTHON_EXTENSION_DISABLE_ACTIVATE_TERMINAL;
+    let i = key.lastIndexOf('.');
+
+    let config = workspace.getConfiguration(key.slice(0, i));
+    await config.update(key.slice(i + 1), value, ConfigurationTarget.Global);
+}
+
+
+export function getProceedwithlongpathsdisabled(): boolean {
+    let key = ROBOCORP_PROCEED_WITH_LONG_PATHS_DISABLED;
+    return get<boolean>(key);
+}
+
+
+export async function setProceedwithlongpathsdisabled(value): Promise<void> {
+    let key = ROBOCORP_PROCEED_WITH_LONG_PATHS_DISABLED;
     let i = key.lastIndexOf('.');
 
     let config = workspace.getConfiguration(key.slice(0, i));