+++ robocorp-code/docs/changelog.md
@@ -1,3 +1,10 @@
+NEXT
+-----------------------------
+
+- RCC was upgraded to `v11.33.2`.
+- No longer showing warnings regarding locks at startup.
+- It's now possible to proceed even with long paths disabled. 
+
 New in 0.41.0 (2022-12-13)
 -----------------------------